Update to support field option errors
I've added my hack code, probably not 100% correct for all field types which will check the processor data error string and if it's got a plain string, it's only a field error, but if it contains an object, it will iterate the object key/value pairs and find the matching option value and add the error message below the relevant option and flag only that option as has-error. the double looping means you'll be able to set multiple field option errors at the same time (if present in the data error return string). (Feature Request #502)