Remove call to rebind_field_bindings on change of  .set-system-variable
This Branch is meant to test properly the effect of such acction.
It fixes #2512 and maybe #2658
I understand this function is supposed to bind the option like system vars, magic tags and variables in the options lists in different places of the Form's panel.
In the precise case of a variable, it will be immediately available in options during creation but it is not when the variable is edited.

I suggest we test this as a PR to fix #2512 and open an issue for variable edition availability in options